[ACC-2058] Untangle dependencies on spring-data-rest
