+++ thunx-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,3 +1,4 @@
 com.contentgrid.thunx.api.autoconfigure.AbacAutoConfiguration
+com.contentgrid.thunx.api.autoconfigure.AbacContextAutoConfiguration
 com.contentgrid.thunx.api.autoconfigure.JwtAbacAutoConfiguration
 com.contentgrid.thunx.gateway.autoconfigure.GatewayAutoConfiguration
