+++ thunx-spring-security/build.gradle
@@ -9,17 +9,16 @@ dependencies {
     compileOnly 'org.projectlombok:lombok'
     annotationProcessor 'org.projectlombok:lombok'
 
-    implementation project(':thunx-encoding-json')
-    implementation project(':thunx-predicates-querydsl')
     implementation project(':thunx-spring-api')
-    implementation project(':spring-data-querydsl-predicate-injector')
+    implementation project(':thunx-model')
 
     implementation 'org.springframework.boot:spring-boot'
-    implementation 'org.springframework.data:spring-data-rest-webmvc'
     implementation 'org.springframework.security:spring-security-oauth2-jose'
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
+
+    testImplementation project(':thunx-encoding-json')
 }
 
 test {
